refactor : 그룹 API 인터페이스 삭제, GroupServiceImpl 을 GroupService 로 변경 & FoodGroup 에 attendance(현재인원) 컬럼 추가로 인한 전반적인 GroupService 코드 수정
